Move config methods to extensions.
