+++ Wikipedia/Code/OnThisDayCollectionViewCell.swift
@@ -4,40 +4,7 @@ import UIKit
 class OnThisDayCollectionViewCell: SideScrollingCollectionViewCell {
 
     let timelineView = OnThisDayTimelineView()
-    
-    @objc(configureWithOnThisDayEvent:dataStore:layoutOnly:shouldAnimateDots:)
-    func configure(with onThisDayEvent: WMFFeedOnThisDayEvent, dataStore: MWKDataStore, layoutOnly: Bool, shouldAnimateDots: Bool) {
-        let previews = onThisDayEvent.articlePreviews ?? []
-        let currentYear = Calendar.current.component(.year, from: Date())
         
-        titleLabel.textColor = .wmf_blue
-        subTitleLabel.textColor = .wmf_customGray
-        
-        titleLabel.text = onThisDayEvent.yearWithEraString()
-
-        if let eventYear = onThisDayEvent.year {
-            let yearsSinceEvent = currentYear - eventYear.intValue
-            subTitleLabel.text = String.localizedStringWithFormat(WMFLocalizedDateFormatStrings.yearsAgo(), yearsSinceEvent)
-        } else {
-            subTitleLabel.text = nil
-        }
-            
-        descriptionLabel.text = onThisDayEvent.text
-        
-        articles = previews.map { (articlePreview) -> CellArticle in
-            return CellArticle(articleURL:articlePreview.articleURL, title: articlePreview.displayTitle, description: articlePreview.descriptionOrSnippet(), imageURL: articlePreview.thumbnailURL)
-        }
-        
-        let articleLanguage = (onThisDayEvent.articlePreviews?.first?.articleURL as NSURL?)?.wmf_language
-        descriptionLabel.accessibilityLanguage = articleLanguage
-        semanticContentAttributeOverride = MWLanguageInfo.semanticContentAttribute(forWMFLanguage: articleLanguage)
-        
-        isImageViewHidden = true
-        timelineView.shouldAnimateDots = shouldAnimateDots
-
-        setNeedsLayout()
-    }
-    
     override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         super.traitCollectionDidChange(previousTraitCollection)
