+++ Wikipedia/Code/OnThisDayExploreCollectionViewCell.swift
@@ -3,13 +3,6 @@ import UIKit
 @objc(WMFOnThisDayExploreCollectionViewCell)
 class OnThisDayExploreCollectionViewCell: OnThisDayCollectionViewCell {
 
-    @objc(configureWithOnThisDayEvent:previousEvent:dataStore:layoutOnly:)
-    func configure(with onThisDayEvent: WMFFeedOnThisDayEvent,  previousEvent: WMFFeedOnThisDayEvent, dataStore: MWKDataStore, layoutOnly: Bool) {
-        bottomTitleLabel.textColor = .wmf_blue
-        bottomTitleLabel.text = previousEvent.yearWithEraString()
-        super.configure(with: onThisDayEvent, dataStore: dataStore, layoutOnly: layoutOnly, shouldAnimateDots: false)
-    }
-
     private lazy var whiteTopAndBottomGradientLayer: CAGradientLayer = {
         let layer = CAGradientLayer()
         let white = UIColor.white.cgColor
